Adapt code also to inherit based testing
For all test classes it will be checked if there are 'setUp' and 'teardown' methods. If not then they will be generated.
This approach ensures that also by failing test the metrics are taken by 'tearDown'.

Unfortunately the test results are not accessible over Junit. So therefore a separate script is necessary which merges the surefire-results into the data set.

Furthermore, the iteration and output file can be passed over the command line.